% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\docType{data}
\name{expr}
\alias{expr}
\title{gene expression demo
A matrix frame storing gene expression data.}
\format{
A matrix with 19092 rows and 100 variables.
}
\usage{
expr
}
\description{
gene expression demo
A matrix frame storing gene expression data.
}
\keyword{datasets}
