% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funRelatedGeneSet.R
\docType{data}
\name{funRelatedGeneSet}
\alias{funRelatedGeneSet}
\title{We collected 17 functional gene sets}
\format{
A GeneSetCollection object.
}
\usage{
funRelatedGeneSet
}
\description{
A GeneSetCollection object containing 17 functional gene sets.
}
\keyword{datasets}
